### 2017-03-06 040dbc3
README: add link to godoc
