### 2017-04-08 #195 @pjotrp
@kortschak I'll look for a reviewer. Since you are an experienced JOSS submitter I would like to ask you to prepare the submission for review. Please look at the tick boxes at the head of one of your other reviews and see if you can tick as much as possible.
