### 2017-04-08 #234 @whedon
[REVIEW]: bíogo/ncbi: interfaces to NCBI services for the Go language
**Submitting author:** @kortschak (<a href="http://orcid.org/0000-0001-8295-2301">Robert Daniel Kortschak</a>)
**Repository:** <a href="https://github.com/biogo/ncbi" target ="_blank">https://github.com/biogo/ncbi</a>
**Version:** v1.0.0

- [x] **Authorship:** Has the submitting author (kortschak) made major contributions to the software?
