+++ README.md
@@ -1,14 +1,14 @@
 ![bíogo](https://raw.githubusercontent.com/biogo/biogo/master/biogo.png)
 
-#NCBI
+# NCBI
 
 [![Build Status](https://travis-ci.org/biogo/ncbi.svg?branch=master)](https://travis-ci.org/biogo/ncbi) [![GoDoc](https://godoc.org/github.com/biogo/ncbi?status.svg)](https://godoc.org/github.com/biogo/ncbi)
 
-##Installation
+## Installation
 
         $ go get github.com/biogo/ncbi/...
 
-##Overview
+## Overview
 
 ncbi provides API interfaces to NCBI services.
 
@@ -20,12 +20,12 @@ ncbi provides API interfaces to NCBI services.
 
 If you use bíogo, please cite Kortschak and Adelson "bíogo: a simple high-performance bioinformatics toolkit for the Go language", doi:[10.1101/005033](http://biorxiv.org/content/early/2014/05/12/005033).
 
-##Library Structure and Coding Style
+## Library Structure and Coding Style
 
 The coding style should be aligned with normal Go idioms as represented in the
 Go core libraries.
 
-##Copyright and License
+## Copyright and License
 
 Copyright ©2011-2013 The bíogo Authors except where otherwise noted. All rights
 reserved. Use of this source code is governed by a BSD-style license that can be
