### 2017-04-10 70ed7a8
README: add links to Entrez and BLAST
