### 2017-04-10 c70234d
README: add links to Go installation
