### 2017-04-10 47df513
README: add community guidelines
