### 2017-04-10 a5a452f
README: add examples
