### 2017-04-10 #234 @kortschak
@ngoto I think this is ready for you to take a look.
