### 2017-04-24 #234 @ngoto
I'm sorry too late.

I'd like to run examples in README.md.
I tried to copy-and-paste code from README.md, saved and builded, but following errors:

```
$ go build  readme-entrez.go
# command-line-arguments
./readme-entrez.go:38: undefined: query
$ go build readme-blast.go
can't load package:
readme-blast.go:1:1: expected 'package', found 'const'
readme-blast.go:3:2: expected ';', found 'IDENT' tool
```
