### 2017-04-24 87f7511
README: fix entrez example and document BLAST as not self contained
