### 2017-04-24 #234 @kortschak
Sorry. I have fixed the entrez example. That was a copy-paste error. The issue with the BLAST example is that it is not intended to be a complete runnable program; providing the harness for all that would detract from the code I have provided. I have clarified this in the preamble to that example.

Please comment at [this PR](https://github.com/biogo/ncbi/pull/5).
