### 2017-05-16 #234 @kortschak
@ngoto I believe I have addressed all of your concerns.
