### 2017-05-16 #234 @ngoto
> Performance: Have any performance claims of the software been confirmed?
It seems that no description about performance (only a phrase "light weight" ?). I suppose the running performance is fully depended on NCBI.
