### 2017-05-16 #234 @kortschak
We make no claims about performance since the vast majority of computational effort is made by the NCBI servers or the local BLAST executable.

I don't believe that syntactic weight constitutes performance, but I guess that could be assessed by comparison of exported function count or SLOC with other similar suites.
