### 2017-05-17 #234 @pjotrp
Thanks @ngoto! Looks like we are pretty much done. There are no real claims on performance, so it is OK not to tick that box. If you are happy we can push for publication.
