### 2017-05-23 #234 @kortschak
@arfon Done.

[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.582354.svg)](https://doi.org/10.5281/zenodo.582354)

`doi:10.5281/zenodo.582354`
