### 2017-05-26 e4146ba
Fixing paper for submission
