+++ paper/paper.bib
@@ -44,7 +44,7 @@
 }
 
 @Article{BioPerl,
-  author    = {Stajich J.E., Block D., Boulez K., Brenner S.E., Chervitz S.A., Dagdigian C., Fuellen G., Gilbert J.G., Korf I., Lapp H., Lehväslaiho H., Matsalla C., Mungall C.J., Osborne B.I., Pocock M.R., Schattner P., Senger M., Stein L.D., Stupka E., Wilkinson M.D. and Birney E.},
+  author    = {Stajich J.E. et al},
   title     = {The Bioperl toolkit: Perl modules for the life sciences},
   journal   = {Genome Research},
   year      = {2002},
@@ -55,7 +55,7 @@
 }
 
 @Article{BioPython,
-  author    = {Cock P.J.A., Antao T., Chang J.T., Chapman B.A., Cox C.J., Dalke A., Friedberg I., Hamelryck T., Kauff F., Wilczynski B. and de Hoon M.J.L.},
+  author    = {Cock P.J.A. et al},
   title     = {Biopython: freely available Python tools for computational molecular biology and bioinformatics},
   journal   = {Bioinformatics},
   year      = {2009},
